requirements.txt
tensorflow 
pillow 
numpy 
sklearn
gradio
pandas
matplotlib
seaborn